Linux port: Fix pytge.so for x86_64 Linux
Key changes:
- Fix static initialization order crashes (gEvalState, DemoGame as pointers)
- Apply MMO Kit patches (Move struct extensions, MissionArea stratosphere)
- Add missing globals (gGamePaused, gPrefGuiOpacity, gClientMissionArea)
- Fix rpgPlayer.cc compile errors (member declarations, variable shadowing)
- Update x86UNIXPlatformState initialization for shared library loading
- Fix GCC warnings and x86_64 type compatibility issues
- Add pytge build target to targets.torque.mk

Tested: pytge.so loads and initializes successfully 10/10 runs

+++ engine/platform/types.posix.h
@@ -6,13 +6,13 @@
 #ifndef _TYPESPOSIX_H_
 #define _TYPESPOSIX_H_
 
+#include <stddef.h>
 
 #define FN_CDECL     ///< Calling convention
 
 // size_t is needed to overload new
-// size_t tends to be OS and compiler specific and may need to 
-// be if/def'ed in the future
-typedef unsigned int  dsize_t;      
+// On 64-bit systems, size_t is 64-bit, so dsize_t must match
+typedef size_t dsize_t;
 
 
 /** Platform dependent file date-time structure.  The defination of this structure
